Rename ciphering to crypto
