+++ Tests/test_caesar_cipher.c
@@ -1,4 +1,4 @@
-#include "ciphering/caesar_cipher.h"
+#include "caesar_cipher.h"
 
 #include "unity.h"
 #include "unity_fixture.h"
